Rename file names for internal modules
To make the file names in internal consistent.
