+++ src/frequenz/sdk/_internal/_singleton_meta.py
