+++ src/frequenz/sdk/actor/power_distributing/_battery_status.py
@@ -19,7 +19,7 @@ from frequenz.api.microgrid.inverter_pb2 import ComponentState as InverterCompon
 from frequenz.channels import Receiver, Sender
 from frequenz.channels.util import Select, Timer
 
-from frequenz.sdk._internal.asyncio import cancel_and_await
+from frequenz.sdk._internal._asyncio import cancel_and_await
 from frequenz.sdk.microgrid import connection_manager
 from frequenz.sdk.microgrid.component import (
     BatteryData,
