+++ src/frequenz/sdk/timeseries/_moving_window.py
@@ -16,7 +16,7 @@ import numpy as np
 from frequenz.channels import Broadcast, Receiver, Sender
 from numpy.typing import ArrayLike
 
-from .._internal.asyncio import cancel_and_await
+from .._internal._asyncio import cancel_and_await
 from ._base_types import UNIX_EPOCH, Sample
 from ._resampling import Resampler, ResamplerConfig
 from ._ringbuffer import OrderedRingBuffer
