+++ tests/utils/component_data_streamer.py
@@ -8,7 +8,7 @@ import asyncio
 from dataclasses import replace
 from datetime import datetime, timezone
 
-from frequenz.sdk._internal.asyncio import cancel_and_await
+from frequenz.sdk._internal._asyncio import cancel_and_await
 from frequenz.sdk.microgrid.component import ComponentData
 
 from .mock_microgrid_client import MockMicrogridClient
